Optionally wrap all of OIIO in a (potentially versioned) namespace.  Top level 'make NAMESPACE=foo' triggers it.  All header files have appropriate ifdefs so that the enclosing namespace will be used if specified, but if not it'll do everything as it always has.
Rather long discussion about why this is helpful below:

Let's say your facility develops several libraries, each maintained by
different groups, that use OIIO in various ways.  Let's call them libA,
libB, libC.  These libraries are also versioned as they are developed so
that particular applications can easily lock down on a particular
version.  Suppose app D links against libA.v3, libB.v8, libC.v5.

OK, everything is going swimmingly, until a bug is discovered in libA,
it's fixed, yielding libA v4.  App D is relinked against A.4, B.8, C.5,
yielding a new version of D.

But *oops* it turns out that between A.3 and A.4, OIIO was updated (in
fact, maybe the update to OIIO is instrumental to fixing the bug in
libA).  Now app D has one library (A.4) that expects expects the new
OIIO, and other libraries (B.8 and C.5) that expect the older OIIO. Wow,
that leads to all sorts of badness, from linkage problems all the way to
extremely subtle runtime bugs.

One solution is to force everybody to use all the same library versions,
all the time.  This means that the bug fix in A requires the groups that
make B and C to cut new versions of their libraries, and potentially for
not only app D but also E and F now need to relink.  Ugh.  A few more
groups, libraries, and apps later, you have a nightmare.

The other solution is that the facility versions OIIO as well, and
furthermore wraps the whole thing in a namespace that contains the
version (i.e. namespace OIIOv14).  Now we've transformed the problem:
libA links against OIIOv14, libB and libC link against OIIOv13.  Since
everything is wrapped in the versioned namespaces, there's no conflict
for app D at all -- it links fine, it runs fine, no bugs.  OK, maybe
under the covers it turns out that it'll have TWO ImageCache's (one
pulled in by libA, the other by libB).  But that's slightly inefficient,
but the program will run correctly, and the development of all the apps
and libs doesn't grind to a halt.


git-svn-id: http://svn.openimageio.org/oiio/trunk@987 31904243-a364-42ac-a3db-9305396518ba