Fixes to avoid possible leaking of tiles.
Make the operator= of IB::Iterator and IB::ConstIterator properly
release any tile it previously held.

In ImageCacheImpl::get_tile, only increment the reference count of the
tile if it's valid and we will return it (NOT if we are going to return
NULL for an invalid tile, that will leak!), and also mark the tile as
"used" so that an app that only accesses tiles via get_tile will have
the proper cache behavior with frequently used tiles staying in cache.



git-svn-id: http://svn.openimageio.org/oiio/trunk@1044 31904243-a364-42ac-a3db-9305396518ba