If the TextureSystem was destroyed before the threads calling it died, the cleanup_perthread_info would end up referencing things still inside the TS, oops.
This change makes it so that the thread-dying destructor just deletes
the per-thread data, and the texturesys-destructor just clears the
per-thread pointers (which we keep track of only for invalidation).  In
the end, everything gets freed this way, but there's no longer a
critical order dependence, as far as I can tell.

I found this rather accidentally, but immediately realized that it was
responsible to some debug mode assertions that Cliff reported a while
back, particularly in cases where TS's were repeatedly created and
destroyed. 



git-svn-id: http://svn.openimageio.org/oiio/trunk@1057 31904243-a364-42ac-a3db-9305396518ba