TextureSystem improvements to cubic B-spline interpolation from Chris Kulla.
A recent change to the cubic B-spline interpolation was meant to address
numerical precision issues wherein filtering a texture with constant texel
values could return filtered values slightly higher than the values.
But the replacement code had its own precision problems where it could be
slightly *less*, which was really disturbing when it would make slightly
negative filtered results of strictly nonnegative texel values!

In this new variant, the B-spline evaluation is done with a quite
different technique based on lerps, which guarantees that the filtered
results can never be outside the range of the source texels underneath
the filter kernel.  And guess what, it's also slightly faster than any
of the other techniques we tried!

This patch has FOUR different B-spline evaluation techniques, controlled
by #ifdef's.  Our plan is to leave it like this in case we discover more
problems or wish to do more benchmarks, then in a subsequent version we
will prune all but the top contender(s).


git-svn-id: http://svn.openimageio.org/oiio/trunk@1179 31904243-a364-42ac-a3db-9305396518ba