ImageCache/TextureSystem threading improvements:
* Move the TS PerThreadInfo into the IC.  Bonus: this means that the
benefits of microcache, etc., now accrue to IC clients who don't use TS.

* Move most of the IC stats and all of the TS statistics into this
per-thread structure, eliminating the needs for them to be atomics.

* Change the locks protecting the file and tile caches to spin locks.

The net effect of all this is approximately a 10-15% speedup overall for
TS-heavy applications (both for single thread mode, as well as when
actually multithreaded).  And for some tests, substantial speedups from
changing to spin locks.



git-svn-id: http://svn.openimageio.org/oiio/trunk@1188 31904243-a364-42ac-a3db-9305396518ba