Maketx -- inquire the input data format (which serves as the default output data format, if -d is not used) BEFORE the read() brings the cache into play and potentially changes the apparent data type.
This solves the following problem: since ImageBuf is backed by
ImageCache, which forces everything to uint8 or float internally, and
maketx was asking for the IB's spec after the cache had a chance to mess
with it, all non-uint8 inputs ended up with float outputs. Example: if
the input file was OpenEXR 'half', the texture output would end up
float, making a much larger file than necessary.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1290 31904243-a364-42ac-a3db-9305396518ba