Make the cmake macro for adding image test suites more powerful.
The macro has been renamed to oiio_add_tests() and rewritten to include some
keyword options: IMAGEDIR allows the test to be added conditional on the
existance of an image directory.  URL allows the location of the test images on
the web to be specified as a prompt to the user.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1291 31904243-a364-42ac-a3db-9305396518ba