Add EXEC_INSTALL_PREFIX to cmake build for control of executable install location.
Patch courtesy of Malcolm Humphreys.  This patch allows the install locations
of the binaries (executables and libraries) to be controlled separately from
the install location of the platform-independent data using the build variable
EXEC_INSTALL_PREFIX to override CMAKE_INSTALL_PREFIX if desired.  Users who
don't care to do this are free to ignore it.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1295 31904243-a364-42ac-a3db-9305396518ba