Bugfix for static initialization order of image cache perthread info mutex.
The static initialization order of ImageCacheImpl::m_perthread_info_mutex and
the shared_image_cache instance was wrong, causing m_perthread_info_mutex to be
used after it was destructed.  (It is used in the ImageCacheImpl destructor.)
The correct order is to declare the static m_perthread_info_mutex first so that
it always outlives the shared image cache instance.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1299 31904243-a364-42ac-a3db-9305396518ba