Shore up the TextureSystem with respect to files in which the image origin is not (0,0), and those in which the pixel data window doesn't match the image window (i.e. textures that are crop windows or that have overscan regions).
In order to ameliorate the noticeable slowdown I saw from the extra
per-sample tests we were doing to make this right (which were wholly
unnecessary for the 99% of textures that don't require it), I added a
per-mip-level "LevelInfo" structure that has such things precomputed.
That makes the TS the same speed as before, and also gives us a place to
make future optimizations if we discover other per-mip-level
computations that we can make once rather than per sample.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1332 31904243-a364-42ac-a3db-9305396518ba