Fix bug in testtex.cpp that caused testsuite/texture-missing to fail.
The test actually passed on first review of the recent change that added
the "missing" texture option (for which that test was created).  Reviewers
suggested a minor change in behavior, and after implementing that for final
commit, I neglected to rerun the test and notice that it no longer matched
the reference image.


git-svn-id: http://svn.openimageio.org/oiio/trunk@1395 31904243-a364-42ac-a3db-9305396518ba