Fix iterator invalidation with unordered_map::insert()
In contrast to older various hash_map implementations, the C++ standard
unordered_map class permits iterators to be invalidated after calling
unordered_map::insert(), but only if this has resulted in a rehash.  This was
causing a segfault when the m_tile_sweep iterator became invalid and was later
used inside ImageCacheImpl::check_max_mem().  The fix is to detect when the
sweep iterator has become invalid and if so reset it to the container end().


git-svn-id: http://svn.openimageio.org/oiio/trunk@1452 31904243-a364-42ac-a3db-9305396518ba