This fixes a rare but painful ImageCache deadlock related to untiled images.
imagecache.cpp lines 745ff and 780ff are the actual solution.  The rest
is additional debugging scaffolding and minor cleanup.

The basic problem was as follows: (1) When reading a tile, we FIRST put
it in the cache, then read the pixels into it, then mark it as "pixels
ready"; and when requesting a tile, we look for it in the cache, and if
it's there, we wait until "pixels ready."  This is how we avoid multiple
threads redundantly reading the tile from disk.  (2) When autotiling is
used and a "tile" is requested from an untiled file, since it has to
read tile_height scanlines to generate that one tile, it actually enters
that whole row of tiles into the cache since they'll likely be needed
anyway (thus avoiding reading the same scanlines multiple times,
separately for each tile that overlaps it).

OK, so these two things conspire against us: 
* thread #1 sees the tile it needs isn't in the cache, so it adds the
  tile (but it's not yet read and marked "ready");
* thread #2 sees a different tile (but in the same row) it needs isn't
  there and adds it to the cache, locks the input mutex, reads the
  scanlines, and starts adding the tiles from that row to the
  cache. Eventually it wants to add the same tile that thread #1 has
  just added, sees that it's already in the cache, so waits for
  "pixels_ready".
* thread #1 meanwhile wants to fill in the pixels of the tile it just
  added, so it tries to acquire the input mutex.

Aha, deadlock!  One thread holds the input mutex but is waiting to get
rights to the tile in the cache, the other thread is holding the tile
but needs the input mutex.  A classic case of trouble from two locks
needed simultaneously, briefly, but just long enough for the dining
philosophers to come home to roost.

The solution is for read_untiled to release the input mutex as soon as
it's done actually reading from the file, but before it tries to put
tiles into the cache.  That gives another thread who has added a tile
the opportunity to read from the file and subsequently mark the tile as
"ready."  In other words, never hold the input mutex while you are
calling add_tile_to_cache.



git-svn-id: http://svn.openimageio.org/oiio/trunk@1631 31904243-a364-42ac-a3db-9305396518ba