Factor the ILMBase and OpenEXR detection into CMake modules.
The modules are a drop-in replacement for the old detection code in
src/cmake/externalpackages.cmake, i.e. they search the same list of directories
and support custom library names.

The modules also try to find and use release and debug versions of the
libraries (i.e. Half.lib for release and Halfd.lib for debug on windows). This
is very useful when using MSVC, as linking to libraries with different versions
of the C++ runtime often leads to segfaults.

Patch by: Edgar Velazquez-Armendariz


git-svn-id: http://svn.openimageio.org/oiio/trunk@1635 31904243-a364-42ac-a3db-9305396518ba