Improve DPX 10 and 12 bit loading (which we convert to uint16 internally). Previously, loading 10 and 12-bit DPX images used a bit-depth conversion that did not fill the output coding space.
For example, say you had a value of 1023 in a 10-bit DPX file, and
requested a 16-bit buffer. With the prior code, you would read a value
of 65472.  With this update, you will read a value of 65535.  If you
request a float buffer, before your value would have been 0.999039.  Now
it will be 1.0.

Image writing is unchanged, 'truncation' is still used.

Patch by: jeremys
Review by: lg


git-svn-id: http://svn.openimageio.org/oiio/trunk@1715 31904243-a364-42ac-a3db-9305396518ba