Fixes compile warnings in public headers. (Tested on gcc44, linux)
The issues fall into 3 types:
* Implicit casts (such as treating an int as a float).  These are 
changed to explicit static_casts.

* Declaration of function parameter names sometimes shadow internal 
members.  This is addressed by adding an underscore (_) suffix to the 
function parameter names as needed.

* Functions implemented in public headers have occasional unused 
parameters.  This has been fixed by commenting out the variable names in 
the declaration, /* unused */.   Or when not appropriate, I've moved the 
implementation from the header to the .cpp file. (Which is preferable in 
some cases anyways).

Review by: lg




git-svn-id: http://svn.openimageio.org/oiio/trunk@1762 31904243-a364-42ac-a3db-9305396518ba