Rename BitsPerSample -> oiio:BitsPerSample
This allows BitsPerSample to be distinguished from other metadata read from the file.  I've also updated the openexr writer to not output metadata starting with
"oiio:".  This fixes the case where if one converted a dpx file to an exr, you would end up with incorrect 'BitsPerSample' metadata in the exr.

git-svn-id: http://svn.openimageio.org/oiio/branches/RB-0.9@1781 31904243-a364-42ac-a3db-9305396518ba