maketx : --filter included in hash, and color+hash metadata always computed
The --hash option is deprecated; the SHA-1 hash is now computed in all cases.
The SHA1 computation was also moved to later in the txmake process, so it also
takes into account upstream image changes, such as resizing. constant color metadata is always emitted (though the image dimenstions are not
shrunk unless explicitly requested, such as with --oiio)