Fix for boost-filesystem v3 on Windows.
Ideally this should use QString::fromWCharArray (fn.filename().c_str())
on Windows to preserve the internal format of boost::filesystem::path.
However the offical builds of Qt for msvc do not use wchar_t as a built-in
type (it is a built-in type by default) causing a linking error.
This will not be fixed until Qt5.