Color processing uses OpenColorIO.
OCIO is now a build-time optional dependency. (Defaulting to ON, if it can be
found - otherwise OFF).   The binary ABI for OIIO is fixed no matter which build
option is used, so the library can be swapped out is color processing is later
determined to be needed.  If OCIO is not enabled, a simple color processing
implementation is included in its place.

(ColorTransfer, and all associated code has been removed).

Modifications include:
- include/color.h
- ImageBufAlgo
- maketx

The location of OpenColorIO can be specified using:
OCIO_PATH: (root of OCIO install)
or OCIO_INCLUDES / OCIO_LIBRARIES if you'd like to specify those dir
indepdendently

Remaining issues to address:
- Clarify use of --unpremult flag in maketx
- In color_generic / color_ocio use iterators instead of get_pixel
- Add ocio support to oiio-tool
- Add in-place, raw conversion to imageio.h
- Add support to the texture system to allow for color conversion at tile-load time