Image Processing Operation: Brightness
I am working in a new branch "brightness"
that branches off from master. For simplicity
reasons I copied the common code for the ROI
class and the function parallel_image. Once
approved I will remove the common code.

I tested the operation for grayscale images.
I am waiting for approval for the RGB algorithm
which requires conversion RGB <-> HSL.

Most of the code is similar to the previous operation
contrast.