Histograms fix 1
1. I fixed few things Chris suggested.

2. It is still debatable how do we output text.
I will ask on mail-list also.

3. I am not sure how to use reres, so I just
commented that part for now.

4. Should we have just one histogram function,
instead of public histogram and histogram_impl
as it is now?

5. I plan to add lot more details to the comments
as I go.

6. Chris I am waiting for your feedback on how
to use capture, so that I can run through it
and figure out how to do optional arguments.

That is all in summary what we are left to deal
with.