Fix for handling of UTF-8 file paths on Windows for OpenEXR files.
For most other fileformats this was already fixed, and I thought this wasn't
possible with the current OpenEXR API, but it turns out it is by implementing
a custom input/output stream class.

The code for these is copied from ImfStdIO.h/cpp which are the classes that
would be used otherwise, so exception handling etc should work identical.