Support for OpenEXR 2.x "multi-part" files (and build support for OpenEXR 2.0).
This basically presents as OIIO "multi-image" files (subimages), but the
hitch is that when writing, OpenEXR needs to know the number of subimages
and all the subimage specs when the file is first opened.  Luckily, we
just added a new open() API call flavor for this purpose, so we use it.