maketx bug fixes for --monochrome-detect and --constant-color-detect
--monochrome-detect did silly things for alpha, now only has an affect
if there is no alpha channel.

--constant-color-detect didn't do anything at all!  I assume that was
a bug, now do the right thing, which is to output a maximum of tile-sized
(usually 64x64) res for the MIPmap, rather than whatever ungodly resolution
somebody had used for the uniformly-colored input texture.

While I'm at it, I'm making the --oiio flag NOT automatically set
--constant-color-detect, on the principle that --oiio should not do anything
that loses information or changes the pixel data or appearance.