Change channel range notation for ImageInput methods to [chbegin,chend).
This patch altered the ImageInput::read_scanlines, read_tiles,
read_native_scanlines, read_native_tiles, read_native_deep_scanlines,
read_native_deep_tiles so that instead of specifying channel subsets as
(firstchan, nchans), they are specified as [chbegin, chend), to match
how spatial extents are done, as well as how channel ranges already were
specified in ROI and ImageBuf.

We hate changing API meanings, but we really think this is better and
more consistent. Note that the two common uses of channel subsets were
firstchan=0,nchans=nchannels (select all channels) and
firstchan=foo,nchans=1, and we have rigged it so that [chbegin,chend)
returns the same channels in both of these cases (in the latter case,
because we retrieve a minimum of 1 channel), so we believe this is
unlikely to break working code in the vast majority of cases.