Allow test output to exclude certain metadata when printing image info.
This uses the --no-metamatch option recently added to oiiotool.
This is really handy for certain tests that involve dumping oiiotool --info,
to simply make it not print the fields that might change from run to run
(like DateTime), or release to realase (like Software) and thus make
spurious testsuite failures.