Texture: make sure wrap mode "black" wins over "fill" value when they conflict.
Wrap mode "black" directs texture lookups outside s,t [0,1] range to
return 0.  The "fill" option gives the value used for channels requested
but which are not in the map.  What happens for channels not in the map,
for lookups outside [0,1] (when "black" wrap mode is used)?

The previous behavior was to use the fill value for missing channels
everywhere (even outside [0,1] when "black" wrap was specified).  But
upon reflection, we believe a better behavior is that those
out-of-st-range lookups should be black in all channels, regardless of
the fill value.