IBA::channel_append() / oiiotool --chappend and oiitool --chnames
channel_append (oiiotool --chappend) takes two input images and produces
an output that is the channels of the two images appended to each other.
For example, appending rgba.exr and z.exr will give you a rgbaz.exr.

oiiotool --chnames lets you rename some or all of the channels of an image.