Allow oiiotool to loop over image sequences.
Image sequences are specified by having filename arguments to oiiotool
use either a numeric wildcard (designated by '#') or an explicit numeric
range (designated such as '1-10#').  In effect, this will execute the
oiiotool command several times, making substitutions to the sequence
arguments in turn.  For example:

    oiiotool big.#.tif --resize 100x100 -o small.#.tif

if there are files big0001.tif, big0002.tif, and big0003.tif, will be the
equivalent of having issued the following sequence of commands:

    oiiotool big.0001.tif --resize 100x100 -o small.0001.tif
    oiiotool big.0002.tif --resize 100x100 -o small.0002.tif
    oiiotool big.0003.tif --resize 100x100 -o small.0003.tif

The following is also equivalent:

    oiiotool big.1-3#.tif --resize 100x100 -o small.1-3#.tif