oiiotool speedup: for not-huge files, force a read of the whole image up front.
oiiotool in particular tends to need all of an image at once, and with
many cores will parallelize across one image operation (say, resize).
If it's a cached-backed image, that means lots of threads are reading
tiles from the same open image file, and they will lock against each
other for control of the file handle.  That severely limits the thread
efficiency of ImageCache, in ways that would not be a problem if the
threads weren't all trying to read from the same file.  We combat this
by forceing a full read of the file into the IB up front, if it's not
too big.  Big files will still fall back to the ImageCache.  This
strategy is not for all apps, but for oiiotool's usage pattern in
particular, it really helps.