ImageBufAlgo::fillholes_pp and oiiotool --fillholes
Fill in the parts of the image with alpha<1 with plausible smooth
interpolant of nearby pixels. This is great for hole filling as well as
extrapolation -- such as if you have a texture atlas with gaps between
pieces and want to fill in the rest of the image so that texture
sampling will filter nicely by not incorrectly drawing in black pixels
beyond the edge of the pieces.