IBA::resample() and oiiotool --resample
This is similar to resize, but instead of properly filtering, it just
samples (either bilinearly or using only the closest pixel).  This is
much lower quality than a proper resize -- when downsampling, you should
expect to alias, and when upsampling, to either replicate pixels or have
obvious linear artifacts -- but it's also up to 10x faster than a resize
of a substantial factor using a good filter like lanczos3.

For high quality, use resize.  But if you want a low-quality thumbnail
as fast as you can possibly get it, this one is what you want.