DPX - add full support of subimages.
DPXInput was mostly right but there were some minor bugs to fix, we
weren't calling the right version of dpx::Reader::ReadBlock to properly
index the right subimage.

DPXOutput needed a pretty extensive overhaul for submages -- it had been
written without subimage support at all.  So needed to tease apart the
header writing stuff that applied to the whole file versus each
individual subimage.  In the end, it was mostly code movement, not a ton
of changes, other than to the high-level logic of dealing with mutiple
subimages in the file.  DPX is one of those formats where you need to
know the information about all subimages when you first open the file
for output.