Bug fixes in libDPX input of multi-image files.
The old code had two entry points for ReadBlock and ReadImage: one that
took the element number, and the other that took a descriptor (that's
DPX spec speak for things like R, G, RGBA, Luma, etc.) and read from the
first, presumably unique, element with that descriptor.  This seems to
made the assumption that if there were multiple subimages in the file,
they would each have unique descriptors. This is obviously not the case,
for example a DPX file may contain subimages for left and right stereo
views, but both may have the same kRGBA descriptor.

Anyway, the "read element" version was implemented by looking up that
element's descriptor, then calling the "read the element with this
descriptor" function. So of course it was wrong.  Thus, any code using
libDPX to read both views of a stereo DPX file would end up reading the
first image twice.

This patch fixes the "read element" to really and truly read the element
(specified by index). And the "read by descriptor", in addition to being
somewhat non-sensical, wasn't needed anywhere, so I just removed it.