Modifications to range compression/expansion.
The coefficients we use to do range compression turn out not to work
well in practice, nor does it match the old tool we use at SPI that
oiiotool is replacing.  Actually, I was unintentionally given
sub-optimal values when I first implemented it in OIIO, when better ones
were being used elsewhere but we didn't realize there were two sets
floating around, oops.

An intersting consequence is that the old coefficients did not alter
values <= 1, so a LDR image would by definition not need the
transformation applied.  The better coefficients, unfortunately, do
compress the range even for values < 1, which is unfortunate, but at the
end of the day the new coefficients do a better job of the intended
purpose.

This patch:

* Alters the coefficients used for the logarithmic<->linear
  transformation used by IBA::rangecompress / rangeexpand.  These work
  better in practice.

* The docs have been altered to no longer say explicitly that values
  < 1 are unchanged, nor that it's a no-op for LDR image formats.

* Remove various code shortcuts that avoided the transformation for LDR
  images or values < 1.  Gotta do it all the time for results to match
  correctly.

* Changes the default for rangecompress/expand to transform each color
  channel independently, rather than use the luma.  The option is still
  there, we're just changing the default to the one that years of
  experience at SPI was preferred.  This applies both to the IBA functions
  themselves, as well as the commands in oiiotool.

Please note that this also applies to the behavior of maketx when
the --hicomp option.